requirements.txt
fastapi
gunicorn
pyzmq
torch
transformers
uvicorn[standard]
